bigint

+++ database/migrations/2023_09_17_184745_create_actividades_table.php
@@ -21,7 +21,7 @@ class CreateActividadesTable extends Migration
             $table->text('descripcion');
             $table->string('portada');
             $table->string('web');
-            $table->integer('id_ciudad')->unsigned();
+            $table->bigInteger('id_ciudad')->unsigned();
             $table->foreign('id_ciudad')->references('id')->on('ciudad');
         });
     }
